pyproject.toml
[project]
name = "vox"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.6,<3.12"
dependencies = [
    "pillow>=11.1.0",
    "pyaudio>=0.2.14",
    "pyautogui>=0.9.54",
    "pydantic>=2.11.3",
    "pydantic-settings>=2.8.1",
    "pynput>=1.8.1",
    "pyperclip>=1.9.0",
    "pystray>=0.19.5",
    "tenacity>=9.1.2",
    "torch>=2.1.0",
    "whisperx>=3.3.1",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]
